noche de lunes

+++ src/main/java/co/edu/utp/isc/gia/turnero/services/CategoryService.java
@@ -111,7 +111,7 @@ public class CategoryService {
     public TurnResponse generateTurn(long categoryId) {
         Turn newTurn = new Turn();
         String nameTurn;
-        int cont;
+        long cont;
        Category category = null;
         
         if (categoryRepository.existsById(categoryId) ) {
